refactor: configContex -> ConfigContext
